% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMF.R
\name{logCpvMF}
\alias{logCpvMF}
\title{Normalization constant of von Mises-Fisher Distribution}
\usage{
logCpvMF(p, eta)
}
\arguments{
\item{p}{is the dimension of the hypersphere.}

\item{eta}{is the intensity parameter.}
}
\description{
log of the Normalization Constant for the von Mises-Fisher Distribution
of dimension \code{p} with intensity parameter \code{eta}.
}
\examples{
logCpvMF(2, 3.1)
}
