% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARDdata.R
\name{sim.network}
\alias{sim.network}
\title{Simulation of the network}
\usage{
sim.network(dnetwork)
}
\arguments{
\item{dnetwork}{is the matrix of link probabilities.}
}
\description{
Generate networks from the network distribution.
}
\examples{
\\donotrun{
N       <- 500 
zeta    <- 1

# Generate the spherical coordinates
z       <- rvMF(N, c(0, 0, 0))

# Genetate the gregariousness
nu      <- rnorm(N, -1.35, 0.37)

# Generale degrees
d       <- runif(N, 0, 45)

# Network distribution
dist    <- sim.dnetwork(nu, d, zeta, z)

# Generate network
G       <- sim.network(dist)}
}
\seealso{
\code{\link{sim.dnetwork}}
}
