% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnetwork.R
\name{fit.dnetwork}
\alias{fit.dnetwork}
\title{Network Distribution}
\usage{
fit.dnetwork(
  object,
  traitARD,
  traitnonARD = NULL,
  m,
  burnin = NULL,
  progress = TRUE
)
}
\arguments{
\item{object}{is an `estim.ARD` object returned by \link{\code{mcmcARD}}}

\item{traitARD}{is the matrix of traits for individuals with ARD. The entry (i, k) is 1 if i has the trait k and 0 otherwise.}

\item{traitnonARD}{is the matrix of traits for individuals without ARD. The entry (j, k) is 1 if j has the trait k and 0 otherwise.}

\item{m}{is the number of neighbours used to compute the gregariousness and the degree for individuals without ARD.}

\item{burnin}{is the number of simulations from the posterior distribution used as burn-in. The network distribution will be computed
used the simulation from the iteration \code{burnin + 1}.}

\item{progress}{is logical; if TRUE, the progression will be printed in the console.}
}
\value{
a matrix of the network distribution
}
\description{
\code{fit.dnetwork} compute the network distribution from the simulations of the posterior distribution of the 
network formation model.
}
\details{
--- VINCENT Any details? ---
}
\examples{
\\donotrun{
FORTCOMING}
}
