% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcSAR.R
\name{mcmcSAR}
\alias{mcmcSAR}
\title{Bayesian Estimator of SAR model}
\usage{
mcmcSAR(
  formula,
  contextual = TRUE,
  start,
  G0 = NULL,
  hyperparms,
  iteration = 2000,
  ctrl.mcmc = list(),
  data
)
}
\description{
Bayesian Estimator of SAR model
}
\seealso{
\code{\link{sim.IV}}
}
